+++ src/python/pants/backend/docker/dependencies_test.py
@@ -5,11 +5,11 @@ from textwrap import dedent
 
 import pytest
 
-from pants.backend.docker import parser
 from pants.backend.docker.dependencies import InjectDockerDependencies, inject_docker_dependencies
+from pants.backend.docker.dockerfile_parser import rules as parser_rules
 from pants.backend.docker.target_types import DockerDependencies, DockerImage
 from pants.backend.python.target_types import PexBinary
-from pants.backend.python.util_rules import pex
+from pants.backend.python.util_rules.pex import rules as pex_rules
 from pants.engine.addresses import Address
 from pants.engine.target import InjectedDependencies
 from pants.testutil.rule_runner import QueryRule, RuleRunner
@@ -19,8 +19,8 @@ from pants.testutil.rule_runner import QueryRule, RuleRunner
 def rule_runner() -> RuleRunner:
     rule_runner = RuleRunner(
         rules=[
-            *parser.rules(),
-            *pex.rules(),
+            *parser_rules(),
+            *pex_rules(),
             inject_docker_dependencies,
             QueryRule(InjectedDependencies, (InjectDockerDependencies,)),
         ],
