+++ src/python/pants/backend/docker/rules.py
@@ -5,7 +5,7 @@ from pants.backend.docker.dependencies import rules as dependencies_rules
 from pants.backend.docker.docker_binary import rules as binary_rules
 from pants.backend.docker.docker_build import rules as build_rules
 from pants.backend.docker.docker_build_context import rules as context_rules
-from pants.backend.docker.parser import rules as parser_rules
+from pants.backend.docker.dockerfile_parser import rules as parser_rules
 
 
 def rules():
